Creation of detection, topshooter, and bottomshooter subsystem
